# src/AppForDevices.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Purchases/GetPurchase": {
      "get": {
        "tags": [
          "Purchases"
        ],
        "operationId": "GetPurchase",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Purchases/CreatePurchase": {
      "post": {
        "tags": [
          "Purchases"
        ],
        "operationId": "CreatePurchase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Rentals/GetRental": {
      "get": {
        "tags": [
          "Rentals"
        ],
        "operationId": "GetRental",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Rentals/CreateRental": {
      "post": {
        "tags": [
          "Rentals"
        ],
        "operationId": "CreateRental",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RentalForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RentalForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RentalForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Repairs/GetRepairsForMaintenance": {
      "get": {
        "tags": [
          "Repairs"
        ],
        "operationId": "GetRepairsForMaintenance",
        "parameters": [
          {
            "name": "repairName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repairScale",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cost_less_or_equal",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepairDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepairDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepairDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaymentMethodTypes": {
        "enum": [
          "CreditCard",
          "PayPal",
          "Cash"
        ],
        "type": "string"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "PurchaseDetailDTO": {
        "required": [
          "deliveryAddress",
          "nameSurname",
          "paymentMethod",
          "userName"
        ],
        "type": "object",
        "properties": {
          "deliveryAddress": {
            "minLength": 1,
            "type": "string",
            "format": "multiline"
          },
          "purchaseItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItemDTO"
            },
            "nullable": true
          },
          "userName": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "nameSurname": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethodTypes"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "totalQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PurchaseForCreateDTO": {
        "required": [
          "deliveryAddress",
          "nameSurname",
          "paymentMethod",
          "userName"
        ],
        "type": "object",
        "properties": {
          "deliveryAddress": {
            "minLength": 1,
            "type": "string",
            "format": "multiline"
          },
          "purchaseItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItemDTO"
            },
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "totalQuantity": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "userName": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "nameSurname": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethodTypes"
          }
        },
        "additionalProperties": false
      },
      "PurchaseItemDTO": {
        "required": [
          "quantity"
        ],
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "priceForPurchase": {
            "type": "number",
            "format": "double"
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QualityType": {
        "enum": [
          "New",
          "Second_han",
          "Reconditioned"
        ],
        "type": "string"
      },
...
      "RentalDetailDTO": {
        "required": [
          "deliveryAddress",
          "name",
          "paymentMethod",
          "surname"
        ],
        "type": "object",
        "properties": {
          "rentalDateFrom": {
            "type": "string",
            "format": "date-time"
          },
          "rentalDateTo": {
            "type": "string",
            "format": "date-time"
          },
          "deliveryAddress": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string",
            "format": "multiline"
          },
          "name": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "surname": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string"
          },
          "rentalItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalItemDTO"
            },
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethodTypes"
          },
          "TotalPrice": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "rentalDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "RentalForCreateDTO": {
        "required": [
          "deliveryAddress",
          "name",
          "paymentMethod",
          "surname"
        ],
        "type": "object",
        "properties": {
          "rentalDateFrom": {
            "type": "string",
            "format": "date-time"
          },
          "rentalDateTo": {
            "type": "string",
            "format": "date-time"
          },
          "deliveryAddress": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string",
            "format": "multiline"
          },
          "name": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "surname": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string"
          },
          "rentalItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalItemDTO"
            },
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethodTypes"
          },
          "TotalPrice": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "RentalItemDTO": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string",
            "nullable": true
          },
          "qualityType": {
            "$ref": "#/components/schemas/QualityType"
          },
          "priceForRental": {
            "type": "number",
            "format": "double"
          },
          "deviceId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RepairDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "format": "double"
          },
          "scale": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}